feat: store 관리 방법론 적용
